Simplify build system and add Property Value Converters for Umbraco 8.1.0
