+++ .gitignore
@@ -1,5 +1,7 @@
+.vs/
+bin/
+obj/
 node_modules/
-dist/
-buildpackage/artifacts/*.zip
-buildpackage/artifacts/*.nupkg
-/.vs/NestingContently/v16
+*.user
+*.min.js
+/dist/App_Plugins/NestingContently
