--- BuildPackage/Build.bat
@@ -1,10 +1,0 @@
-ECHO APPVEYOR_REPO_BRANCH: %APPVEYOR_REPO_BRANCH%
-ECHO APPVEYOR_REPO_TAG: %APPVEYOR_REPO_TAG%
-ECHO APPVEYOR_BUILD_NUMBER : %APPVEYOR_BUILD_NUMBER%
-ECHO APPVEYOR_BUILD_VERSION : %APPVEYOR_BUILD_VERSION%
-ECHO PACKAGE_SUFFIX : %UMBRACO_PACKAGE_PRERELEASE_SUFFIX%
-cd ..\
-Call npm install
-Call gulp --production --buildversion=%APPVEYOR_BUILD_VERSION% --buildbranch=%APPVEYOR_REPO_BRANCH% --packagesuffix=%UMBRACO_PACKAGE_PRERELEASE_SUFFIX%
-cd BuildPackage
-Call "%programfiles(x86)%\Microsoft Visual Studio\2017\Community\MSBuild\15.0\Bin\MSBuild.exe" package.proj
