--- BuildPackage/Package.xml
@@ -1,36 +1,0 @@
-﻿<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<umbPackage>
-  <info>
-    <package>
-      <name></name>
-      <version>0.0.0</version>
-      <license url=""></license>
-      <url></url>
-      <requirements>
-        <major>0</major>
-        <minor>0</minor>
-        <patch>0</patch>
-      </requirements>
-    </package>
-    <author>
-      <name></name>
-      <website></website>
-    </author>
-    <readme><![CDATA[]]></readme>
-  </info>
-
-  <DocumentTypes />
-  <Templates />
-  <Stylesheets />
-  <Macros />
-  <DictionaryItems />
-  <Languages />
-  <DataTypes />
-
-  <Actions/>
-
-  <control />
-
-  <files />
-
-</umbPackage>
