--- BuildPackage/Tools/MSBuildNugetTasks/MSBuild.NuGet.Tasks.Targets
@@ -1,12 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-	
-	<PropertyGroup>
-		<MSBuildNuGetTasksPath Condition="'$(MSBuildNuGetTasksPath)' == ''">$(MSBuildProjectDirectory)\MSBuildTasks</MSBuildNuGetTasksPath>
-		<MSBuildNuGetTasksLib>$(MSBuildNuGetTasksPath)\MSBuild.NuGet.Tasks.dll</MSBuildNuGetTasksLib>
-	</PropertyGroup>
-	
-	<UsingTask AssemblyFile="$(MSBuildNuGetTasksLib)" TaskName="MSBuild.NuGet.Tasks.ManifestUpdate" />
-	<UsingTask AssemblyFile="$(MSBuildNuGetTasksLib)" TaskName="MSBuild.NuGet.Tasks.Pack" />
-
-</Project>
