+++ NestingContently.Umbraco.ValueConverters/package.nuspec
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<package xmlns="http://schemas.microsoft.com/packaging/2011/08/nuspec.xsd">
+	<metadata>
+		<id>$id$</id>
+		<version>$version$</version>
+		<title>$title$</title>
+		<description>$description$</description>
+		<authors>$authors$</authors>
+		<projectUrl>$projectUrl$</projectUrl>
+		<license type="expression">$license$</license>
+		<dependencies>
+			<group targetFramework=".NETFramework4.7.2">
+				<dependency id="UmbracoCms.Core" version="[8.1,9)" />
+				<dependency id="UmbracoCms.Web" version="[8.1,9)" />
+			</group>
+		</dependencies>
+	</metadata>
+	<files>
+		<file src="bin\$configuration$\net472\*" target="lib\net472" />
+	</files>
+</package>
