+++ NestingContently.Umbraco/App_Plugins/NestingContently/backoffice/editor.html
@@ -1,0 +1,1 @@
+<div ng-controller="nestingContentlyController as vm"></div>
