+++ NestingContently.Umbraco/App_Plugins/NestingContently/package.manifest
@@ -1,0 +1,19 @@
+﻿{
+	propertyEditors: [
+		{
+			alias: "NestingContently",
+			name: "Nesting Contently",
+			icon: "icon-bird",
+			editor: {
+				view: "~/App_Plugins/NestingContently/backoffice/editor.html",
+				valueType: "INT"
+			}
+		}
+	],
+	css: [
+		"~/App_Plugins/NestingContently/backoffice/styles.css"
+	],
+	javascript: [
+		"~/App_Plugins/NestingContently/backoffice/nesting-contently.min.js"
+	]
+}
