+++ NestingContently.Umbraco/NestingContently.Umbraco.csproj
@@ -1,0 +1,36 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+	<PropertyGroup>
+		<TargetFramework>net472</TargetFramework>
+		<Id>NestingContently.Umbraco</Id>
+		<Version>2.1.3</Version>
+		<Title>Nesting Contently</Title>
+		<Description>A property editor for toggling the display state of Nested Content</Description>
+		<Authors>Nathan Woulfe</Authors>
+		<PackageProjectUrl>https://github.com/nathanwoulfe/nestingcontently</PackageProjectUrl>
+		<PackageLicenseExpression>MIT</PackageLicenseExpression>
+		<EnableDefaultItems>false</EnableDefaultItems>
+		<GeneratePackageOnBuild Condition="'$(Configuration)' == 'Release'">true</GeneratePackageOnBuild>
+		<NuspecFile>package.nuspec</NuspecFile>
+		<NuspecProperties>id=$(Id);version=$(Version);title=$(Title);description=$(Description);authors=$(Authors);projectUrl=$(PackageProjectUrl);license=$(PackageLicenseExpression);configuration=$(Configuration)</NuspecProperties>
+	</PropertyGroup>
+	<Target Name="NpmBuild" AfterTargets="BeforeBuild" Condition="'$(Configuration)' != 'Release'">
+		<Exec Command="npm run build" />
+	</Target>
+	<Target Name="NpmPublish" AfterTargets="BeforeBuild" Condition="'$(Configuration)' == 'Release'">
+		<Exec Command="npm ci --prefer-offline" />
+		<Exec Command="npm run publish" />
+	</Target>
+	<Target Name="UmbPack" AfterTargets="Pack">
+		<Exec Command="umbpack pack package.xml -v $(Version) -o $(PackageOutputPath)" />
+	</Target>
+	<ItemGroup>
+		<Content Include="App_Plugins\NestingContently\**" />
+		<None Include="src\**" />
+		<None Include="package.json" />
+		<None Include="package-lock.json">
+			<DependentUpon>package.json</DependentUpon>
+		</None>
+		<None Include="package.nuspec" />
+		<None Include="package.xml" />
+	</ItemGroup>
+</Project>
