+++ NestingContently.Umbraco/package.json
@@ -1,0 +1,24 @@
+{
+	"name": "nestingcontently.umbraco",
+	"version": "1.0.0",
+	"private": true,
+	"scripts": {
+		"build": "rollup src/entry.js --file App_Plugins/NestingContently/backoffice/nesting-contently.min.js --format iife --plugin babel={babelHelpers:'bundled'}",
+		"publish": "rollup src/entry.js --file App_Plugins/NestingContently/backoffice/nesting-contently.min.js --format iife --plugin babel={babelHelpers:'bundled'} --plugin terser"
+	},
+	"babel": {
+		"presets": [
+			"@babel/env"
+		]
+	},
+	"browserslist": [
+		"defaults"
+	],
+	"devDependencies": {
+		"@babel/core": "^7.11.6",
+		"@babel/preset-env": "^7.11.5",
+		"@rollup/plugin-babel": "^5.2.1",
+		"rollup": "^2.28.2",
+		"rollup-plugin-terser": "^7.0.2"
+	}
+}
