+++ NestingContently.Umbraco/src/editor.controller.js
@@ -1,0 +1,17 @@
+﻿// for this to work, we need a property editor, we don't, however, need it to be visible or edited...
+angular.module('umbraco').controller('nestingContentlyController', ['$element', ($element) => {
+	const findAncestor = (el, cls) => {
+		while ((el = el.parentElement) && !el.classList.contains(cls));
+		return el;
+	}
+
+	const propElm = findAncestor($element[0], 'umb-property');
+
+	if (propElm) {
+		var parentElem = propElm.parentElement;
+		if (parentElem.classList.contains('umb-nested-content-property-container')) {
+			parentElem.style.display = 'none';
+		}
+	}
+}
+]);
