+++ NestingContently.Umbraco/src/entry.js
@@ -1,0 +1,3 @@
+import "./app";
+import "./button.component";
+import "./editor.controller";
