+++ NestingContently.sln
@@ -1,0 +1,38 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30503.244
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{BB056A3C-F8FF-40DD-BC0C-8A76D99D8513}"
+	ProjectSection(SolutionItems) = preProject
+		.gitignore = .gitignore
+		appveyor.yml = appveyor.yml
+		README.md = README.md
+	EndProjectSection
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NestingContently.Umbraco.ValueConverters", "NestingContently.Umbraco.ValueConverters\NestingContently.Umbraco.ValueConverters.csproj", "{D132F991-BA23-43DC-B826-360C1B3DD43F}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NestingContently.Umbraco", "NestingContently.Umbraco\NestingContently.Umbraco.csproj", "{73644C7C-CC4D-43F1-AAA7-B5A7810A1501}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D132F991-BA23-43DC-B826-360C1B3DD43F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D132F991-BA23-43DC-B826-360C1B3DD43F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D132F991-BA23-43DC-B826-360C1B3DD43F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D132F991-BA23-43DC-B826-360C1B3DD43F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{73644C7C-CC4D-43F1-AAA7-B5A7810A1501}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{73644C7C-CC4D-43F1-AAA7-B5A7810A1501}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{73644C7C-CC4D-43F1-AAA7-B5A7810A1501}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{73644C7C-CC4D-43F1-AAA7-B5A7810A1501}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E69BCDDF-100C-4178-903A-FEC3DBDB6EA2}
+	EndGlobalSection
+EndGlobal
